requirements.txt
# GasFill Python Backend Requirements
# Core FastAPI framework and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Authentication and security
pyjwt==2.8.0
python-multipart==0.0.6

# Data validation and serialization
pydantic[email]==2.5.0

# CORS middleware (included in FastAPI but explicit for clarity)
# fastapi already includes starlette with CORS

# Optional: Database support (uncomment as needed)
# sqlalchemy==2.0.23
# psycopg2-binary==2.9.9  # PostgreSQL
# pymongo==4.6.0  # MongoDB

# Optional: WebSocket support
websockets==12.0

# Development dependencies
# pytest==7.4.3
# pytest-asyncio==0.21.1
# httpx==0.25.2  # For testing

# Production dependencies
# gunicorn==21.2.0  # Production WSGI server
# python-dotenv==1.0.0  # Environment variables